fix(poa): insert block via blockchain tree engine before sending FCU message
The engine will look for existing block indecies before making the canon
chain. If that block hash does not exist it will assume there is a gap
and reach out to peers. Since we are creating that block peers will not
have it either ending up in a endless sync.