botanix-lib: Optimize mint contract validation
Several things were off.
It was expecting the commitment as the last element; this doesn't work,
the user needs to add some leeway in order to not have to redo the tx in
case a bitcoin block gets added.
On the other hand, the header of the block containing the pegin in any
position, but that one can (and should) be fixed to being the first in
the list.